+++ examples/file_storage/Log.h
@@ -9,11 +9,11 @@
 #define LOG_STATIC_BUFFER_SIZE 256
 #define LOG_STREAM Serial
 
-// Storage settings
-#define LOG_STORAGE_ENABLE 1             // MUST BE ENABLED
-#define LOG_STORAGE_LEVEL LOG_LEVEL_WARN // Store WARN and ERROR messages
-#define LOG_STORAGE_MAX_BUFFER_SIZE 4096
-#define LOG_STORAGE_MAX_FILE_SIZE (1024 * 100) // 100KB per file
-#define LOG_STORAGE_MAX_FILES 3
+// File storage settings
+#define LOG_FILE_ENABLE 1             // MUST BE ENABLED
+#define LOG_FILE_LEVEL LOG_LEVEL_WARN // Store WARN and ERROR messages
+#define LOG_FILE_MAX_BUFFER_SIZE 4096
+#define LOG_FILE_MAX_SIZE (1024 * 100) // 100KB per file
+#define LOG_FILE_MAX_FILES 3
 
 #include <FormatLog.h>